+++ changelog.md
@@ -1,10 +1,18 @@
 # CHANGELOG
 
+## [0.1.4] 2020-12-01
 
 ## [0.1.2] 2020-06-22
 
 ### Features and Fixes
+
+#### 0.1.4
+
+- Fix use of ruler as the 0.7.7 API appears to be
+- Don't need to hold ALT down now as Ruler is locked until confirmation received
+
 #### 0.1.2
+
 - GM doesn't have to approve their own movement
 - Option to disable request dialog, automatically send request to GM